Incorrect instance variables have changed into the corrected ones

+++ src/main/java/org/example/Healthplan.java
@@ -2,17 +2,17 @@ package org.example;
 import org.example.enums.Plan;
 
 public class Healthplan {
-    private long id;
+    private int id;
     private String name;
     private Plan plan;
 
-    public Healthplan(long id, String name, Plan plan) { // Constructor is used for setting three instance variables!
+    public Healthplan(int id, String name, Plan plan) { // Constructor is used for setting three instance variables!
         this.id = id;
         this.name = name;
         this.plan = plan;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
